+++ .github/workflows/docker_image_push.yml
@@ -40,8 +40,8 @@ jobs:
         pytest ./tests/unit_tests --cov=./ --cov-append 
     - name: Run Miners
       run: python tests/miners/text/test_gpt2_exodus.py --cov=./ --cov-append 
-    - name: Run dataloader tests
-      run: pytest ./tests/integration_tests/test_dataloader_ipfs.py --cov=./ --cov-append 
+    - name: Run dataset tests
+      run: pytest ./tests/integration_tests/test_dataset_ipfs.py --cov=./ --cov-append 
     - name: Code Coverage
       run: codecov
     - name: Login to DockerHub
